controller/requirements-dev.txt
mpy-utils
mpremote
micropython-esp32-stubs
esptool